bug #234 clear error occurred after ttl or origin directive
git-svn-id: file:///svn/nsd/trunk@2859 a26ef69c-88ff-0310-839f-98b793d9c207

+++ util.h
@@ -272,7 +272,7 @@ timeval_to_timespec(struct timespec *left,
  * XXX This functions does not check the range.
  *
  */
-long strtottl(const char *nptr, const char **endptr);
+uint32_t strtottl(const char *nptr, const char **endptr);
 
 /*
  * Convert binary data to a string of hexadecimal characters.
